// fe/src/lib/api/generated/model/gTSFirstJobSourceValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type GTSFirstJobSourceValue = typeof GTSFirstJobSourceValue[keyof typeof GTSFirstJobSourceValue];


export const GTSFirstJobSourceValue = {
  ADVERTISEMENT: 'ADVERTISEMENT',
  WALK_IN: 'WALK_IN',
  RECOMMENDED: 'RECOMMENDED',
  FRIENDS: 'FRIENDS',
  SCHOOL_PLACEMENT: 'SCHOOL_PLACEMENT',
  FAMILY_BUSINESS: 'FAMILY_BUSINESS',
  JOB_FAIR_PESO: 'JOB_FAIR_PESO',
  OTHER: 'OTHER',
} as const;